requirements.txt
django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
pillow==10.4.0
python-decouple==3.8
